app/requirements.txt
django
django-environ
mod_wsgi
psycopg2
djangorestframework
djangorestframework-simplejwt
drf-yasg


